Added Neural Network Minimum scaling in a place such that it properly finds the minimum for large scaling
See previous commit. Essentially added the same functions, but for the NN. These were added in a brute-force way because of the difficulty in how the scaling is structured and time constraints.  Perhaps it would be good to revisit a different way of doing this in the future.

Comments/docstrings to come soon